Remove 'inputs' folder under atoms and move all components directly under atoms/
