+++ packages/agentflow/src/atoms/index.ts
@@ -1,13 +1,13 @@
 // UI Components - Internal design system
 export { ArrayInput, type ArrayInputProps } from './ArrayInput'
+export { CodeInput, type CodeInputProps } from './CodeInput'
 export { ConditionBuilder, type ConditionBuilderProps } from './ConditionBuilder'
 export { ExpandTextDialog, type ExpandTextDialogProps } from './ExpandTextDialog'
-export { CodeInput, type CodeInputProps } from './inputs/CodeInput'
-export { JsonInput, type JsonInputProps } from './inputs/JsonInput'
-export { SelectVariable, type SelectVariableProps, type VariableItem } from './inputs/SelectVariable'
+export { JsonInput, type JsonInputProps } from './JsonInput'
 export { MainCard, type MainCardProps } from './MainCard'
 export { type MessageEntry, MessagesInput, type MessagesInputProps } from './MessagesInput'
 export { type AsyncInputProps, type ConfigInputComponentProps, NodeInputHandler } from './NodeInputHandler'
+export { SelectVariable, type SelectVariableProps, type VariableItem } from './SelectVariable'
 // RichTextEditor is exported from the .lazy wrapper (not the real module) to avoid
 // eagerly pulling TipTap + highlight.js into the main bundle. Importing directly
 // from ./RichTextEditor would defeat code-splitting since barrel imports are resolved eagerly.
